Extract result panel to separate component
in order to reuse it in the search view later on
